--- src/components/Footer/Footer.css
@@ -1,19 +1,0 @@
-footer {
-  -webkit-box-sizing: border-box;
-          box-sizing: border-box;
-  background-color: #EDAE49;
-  margin: 0;
-  padding: 1em;
-}
-footer a {
-  color: black;
-}
-footer hr {
-  border: 0.5px solid black;
-}
-
-@media screen and (max-width: 768px) {
-  footer {
-    text-align: center;
-  }
-}
