--- src/components/NavigationBar/NavigationBar.scss
@@ -1,122 +1,0 @@
-.nav {
-  width: 100%;
-  background-color: #30638E;
-  display: flex;
-  justify-content: space-between;
-  height: 20vh;
-
-  * {
-    transition: 150ms ease-in-out;
-  }
-
-  &-logo {
-    color: white;
-    cursor: pointer;
-    padding: 1em 2em;
-    text-decoration: none;
-    transition: 150ms ease-in-out;
-
-    &-big {
-      font-size: 2em;
-      letter-spacing: 0.5rem;
-      text-transform: uppercase;
-    }
-
-    &-small {
-      letter-spacing: 0.25rem;
-      text-align: center;
-      display: flex;
-      align-items: center;
-    }
-
-    &:hover {
-      color: rgb(227, 225, 225);
-    }
-  }
-
-  &-menu {
-    list-style: none;
-    padding: 1em;
-    margin: 0;
-    display: none;
-    flex-direction: row;
-    align-items: flex-end;
-    justify-content: center;
-
-    a {
-      border: 1px solid white;
-      padding: 0.25em 0.5em;
-      color: white;
-      border-radius: 8px;
-      transition: 150ms ease-in-out;
-      margin: 0 0.5em;
-      cursor: pointer;
-      text-decoration: none;
-
-      &:hover {
-        background-color: white;
-        color: #30638E;
-      }
-    }
-  }
-
-  &-dropdown {
-    position: absolute;
-    left: 0;
-    top: 20vh;
-    background-color: #30638E;
-    margin: 0;
-    padding: 1em;
-    list-style: none;
-    width: 100%;
-    text-align: center;
-    display: flex;
-    flex-direction: column;
-
-    a {
-      padding: 1em 0;
-      width: 100%;
-      text-decoration: none;
-      color: white;
-      font-size: 1.08em;
-      text-transform: uppercase;
-
-
-      &:hover {
-        background-color: white;
-        color: #30638E;
-      }
-    }
-  }
-}
-
-.toggle-dropdown {
-  height: 44px;
-  width: 44px;
-  padding: 0.25em;
-  margin: auto 10%;
-  color: white;
-  transition: 50ms ease-in-out;
-  border-radius: 8px;
-  cursor: pointer;
-
-  &:hover {
-    outline: 1px solid white;
-    transform: scale(1.08);
-  }
-
-}
-
-
-
-@media screen and (min-width: 768px) {
-
-  .nav-dropdown,
-  .toggle-dropdown {
-    display: none;
-  }
-
-  .nav-menu {
-    display: flex;
-  }
-}
